+++ pumpkin/src/plugin/api/events/entity/death.rs
