+++ pumpkin/src/world/mod.rs
@@ -2702,13 +2702,13 @@ impl World {
     pub async fn spawn_entity(&self, entity: Arc<dyn EntityBase>) {
         let base_entity = entity.get_entity();
         if let Some(server) = self.server.upgrade() {
-            let event = crate::plugin::entity::entity_spawn::EntitySpawnEvent::new(
+            let event = crate::plugin::entity::spawn::EntitySpawnEvent::new(
                 base_entity.entity_uuid,
                 base_entity.entity_type,
             );
             let event = server
                 .plugin_manager
-                .fire::<crate::plugin::entity::entity_spawn::EntitySpawnEvent>(event)
+                .fire::<crate::plugin::entity::spawn::EntitySpawnEvent>(event)
                 .await;
             if event.cancelled {
                 return;
